refactor(tidbtype): move TiDB metadata definitions to dedicated package
- Extract TiDB-specific constants and functions from `common` package to new `tidbtype` package
- Create separate files for classic and nextgen builds to handle different TiDB versions
- Update all imports to use new package location for `JobTableID`, `JobHistoryID`, and `ValidateKeyspace`
- Maintain backward compatibility through package restructuring
