+++ pkg/common/event/mounter.go
@@ -24,6 +24,7 @@ import (
 	"github.com/pingcap/ticdc/pkg/common"
 	"github.com/pingcap/ticdc/pkg/errors"
 	"github.com/pingcap/ticdc/pkg/integrity"
+	"github.com/pingcap/ticdc/pkg/tidbtype"
 	"github.com/pingcap/tidb/pkg/meta/model"
 	"github.com/pingcap/tidb/pkg/tablecodec"
 	"github.com/pingcap/tidb/pkg/types"
@@ -199,7 +200,7 @@ func ParseDDLJob(rawKV *common.RawKVEntry, ddlTableInfo *DDLTableInfo) (*model.J
 	tableID := tablecodec.DecodeTableID(rawKV.Key)
 
 	// parse it with tidb_ddl_job
-	if tableID == common.JobTableID {
+	if tableID == tidbtype.JobTableID {
 		row, err := decodeRow(rawKV.Value, recordID, ddlTableInfo.DDLJobTable, time.UTC)
 		if err != nil {
 			return nil, errors.Trace(err)
@@ -208,7 +209,7 @@ func ParseDDLJob(rawKV *common.RawKVEntry, ddlTableInfo *DDLTableInfo) (*model.J
 		v = datum.GetBytes()
 
 		return parseJob(v, rawKV.StartTs, rawKV.CRTs, false)
-	} else if tableID == common.JobHistoryID {
+	} else if tableID == tidbtype.JobHistoryID {
 		// parse it with tidb_ddl_history
 		row, err := decodeRow(rawKV.Value, recordID, ddlTableInfo.DDLHistoryTable, time.UTC)
 		if err != nil {
