+++ pkg/tidbtype/tidb_metadef_nextgen.go
@@ -1,0 +1,27 @@
+// Copyright 2025 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build nextgen
+
+package tidbtype
+
+import (
+	"github.com/pingcap/tidb/pkg/meta/metadef"
+)
+
+const (
+	// JobTableID is the id of `tidb_ddl_job`.
+	JobTableID = metadef.TiDBDDLJobTableID
+	// JobHistoryID is the id of `tidb_ddl_history`
+	JobHistoryID = metadef.TiDBDDLHistoryTableID
+)
